refactor: Split FastVizSwitcher into multiple files for readability (#30662)
